requirements.txt
requests
pyfiglet
pysocks
